+++ od-win32/dxwrap.cpp
@@ -1208,7 +1208,7 @@ int DirectDraw_Start (void)
 	guid = NULL;
 	if (isfullscreen ()) {
 		MultiDisplay *md = getdisplay(&currprefs, 0);
-		int disp = md - Displays;
+		int disp = (int)(md - Displays);
 		if (disp < 0)
 			disp = 0;
 		if (disp >= MAX_DISPLAYS)
